backend/requirements.txt
Django==3.2
djangorestframework==3.12
django-filter==23.1
djoser==2.2
urllib3==1.26
Pillow==9.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
